Clarification on X.FL connectors
Add part number for the X.FL connector to be soldered on the copper board.

I might be confusing the two. I was lost a bit by the use of "X.FL sockets" and "X.FL connectors". I was expecting both sides to be "connectors" and the socket to be the receptacle part (R-SMT-1) but it seems to refer to the plug (PR-SMT1-2). It does not help that hirose calls it "plug receptacle" as opposed to the "receptacle". 

If I understood correctly, I would suggest to use "plug" or "plug receptacle" instead of socket on this page.